Apply VW data in time order (#6324)
## Describe your changes

With some surgery coming up for VW snpashot_seqno, this refactors to
make our future snapshot_seqno change easier.

Historically, we would segment data as "portable" vs "speculative" in
the VW (each of those would then be segmented as "vault data" or
"document data"). It was written this way likely to make the code in
`portablize_data.rs` easier at the time of writing. But we can just
build this portable/speculative view there when it's needed instead of
subjecting the whole VW to use it

When the VW would serve a request for a read (ie `vw.get(di)`), we chose
to first serve speculative data, then portable data. This was fine when
a VW would never have two pieces of data with the same DI. But, it is in
fact possible that a tenant has speculative data that overwrites
previously written portable data by another tenant.

So, instead of proxying the DI -> speculative/portable -> document /
vault data at read time, we pre-assemble a map of DI -> data and store
this on the VaultWrapper. The key here is that we apply data into this
map in chronological order of when it became _visible_ to the vault
wrapper.

Now, there's some interesting logic to determine when a piece of data
became "visible" to a VW:
- If the piece of data was added by the SV, it was "visible" as soon as
it was added
- If the piece of data was added by another SV (aka another tenant), it
was "visible" when it was portablized

## Caveats


## Test Plan

The intent is that this changes pretty no existing behavior in a
meaningful way - we had previously chosen to always show speculative
data over portable data when there's a clash, and this generally has the
same effect for today's use cases.

The VW has lots of good coverage, so tests should pass.

Added a new test for applying data in chronological order