[FP-6448] Force create a workflow for auth tokens created via API (#6397)
## Describe your changes

Kind of similar to the redo flows - when an auth token is created for a
user specifically linked to a playbook, we always want a new workflow to
be created.

This allows a user to re-onboard onto an ob config they've already
onboarded onto

## Issue: <FP-???>

## Caveats

This specifically _doesn't_ set `is_redo` on the workflows. I'm not sure
if we really even read this anymore. It seems to have pretty reasonable
behavior though

## Test Plan

Integration test, tested locally.