[FP-6433] Issue new token when stepping up in /identify/verify (#6419)
## Describe your changes

when we build auth and footprint.js starts to take in an authToken
passed in by the tenant, we want to make sure that the authToken used by
the tenant doesn’t get these new permissions. it could be bad if
- tenant generates an auth token for a user. this auth token doesn’t
have any permissions - it simply acts as an identifier
- using this auth token, the user verifies their phone number
- the auth token that’s under the tenant’s possession now has
permissions to edit the user’s data! scary

this change to POST /hosted/identify/verify starts returning a new auth
token in these step up cases. we’ll have to update the frontend to also
replace the auth token it’s using with the new one issued by POST
/hosted/identify/verify

## Caveats

Will stop mutating the old token once the frontend migrates to use these
new tokens

## Test Plan

tests