Show docs in the same order as they are rendered in the navigation sidebar (#6465)
## Describe your changes

We had different code that grouped articles into sections than the code
that assembled the list of all articles. This unifies the code used to
assemble all articles and the code used to build the navigation
sections. Now, when the order of the articles changes, the order of the
navigation sections will change too

## Test Plan

Locally. This shows scrolling through the sections highlights the
navigation in the same order:

https://github.com/onefootprint/monorepo/assets/21066341/65661671-fc6b-4ca4-97db-73a1ddb96e2d