[FP-6472] Support trigger for non-portable vaults (#6472)
## Describe your changes

Now that we don't use authenticated tokens (just identified), we can
send these tokens to unverified contact info on a non-portable vault.
This is all made possible with the new user-specific onboarding feature
and the addition of unauthed tokens

This option won't be displayed in the dashboard for users that haven't
already onboarded onto a playbook. And if this is attempted to be used
for a non-portable user who hasn't onboarded onto any playbook, it will
error anyways since there's no playbook to use for the redo flow

## Test Plan

Same flow as identified tokens. Tested locally, but existing test
coverage is good