[FP-6456] Semi-gracefully handle when multiple vaults are identified by fingerprint (#6474)
## Describe your changes

This chooses pretty arbitrary fallback behavior for the case when we
loate multiple vaults that match a single fingerprint. We could probably
get more clever...

As we start allowing vaults created via API to become portablized, it's
no longer the case that only one vault will have a global fingerprint
for a given phone number. and, it was already possible for multiple
vaults to have a global fingerprint for a given email.

## Test Plan

Existing test