Add integration test for inherited auth (#6564)
## Describe your changes

Bringing this back in a separate PR since it was tricky to get right...

I want to be able to have some test that we're not always allowing a
tenant to inherit auth - the auth event has to be recent enough.

But, the ci that runs integration tests doesn't always have historical
data. So, we just skip the test in that case

In the future, i think we'll want to have some logic in the auth token
extractor that checks the time at which the auth event occurred for
expiry too. This would help protect against rogue issuing of auth tokens