Some fixes to the black image issue (#6605)
## Describe your changes
Context on slack:
[link](https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1698704952354559)

We are not 100% certain as to what's going on, however, I believe it has
to do with drawing wrong pixels on the canvas.

- Possibility 1 (also as a fix to the canvasContext2D null issues): We
are reading the canvas too frequently on the autocapture hook. While the
theory is that reading multiple **types** of context from the same
canvas may lead to null context return, it's also possible that multiple
reading may cause issues. So, I separated the autocapture canvas from
the image drawing canvas.
- Possibility 2: User might be clicking capture when the video is still
not ready to be captured. There are cases on internet where the video
first frame comes as black. So, disabling the capture button until the
video is ready (what defines the video as "ready" is tricky and debated
tho).
- Possibility 3: Wrongly calculating the dimensions. We may be wrong
calculating the desired image width and height. So, we are capturing
whole image for the selfie (since portrait is also recommended by
Incode) and adding logging for wrong desired dimensions calculation.


## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->