Treat AlpacaKycState::PendingReview as a completed WF state (#6684)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

At the moment, the AlpacaKyc workflow has a specific PendingReview state
for users who are put into review. When the review is completed, the
workflow then transitions to Complete.
We've recently added the concept of `workflow.completed_at` and various
bits of logic that are contingent on that. One side effect here is that
a tenant will get an error if they try to trigger redo-kyc while the
user is in the PendingReview alpaca state.

https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1698954281196289

This PR just makes it so we set `workflow.completed_at` when we
transition an AlpacaKyc workflow to PendingReview. A minor quirk here is
we would actually be re-setting `completed_at` when/if the user actually
completes the review. I think this isn't a big deal.

All this is ideally temporary as I blast through this stack which
sunsets the bespoke AlpacaKyc workflow for good:
https://github.com/onefootprint/monorepo/pull/6632

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->