Prevent infinite stepup loopage (#6640)
## Describe your changes
Introduces a concept of allowed_actions for a rule execution which
stipulates which kinds of actions are eligible. Once we have already
step'd up (or collected doc at all),
StepUp is no longer a possible action we can take. So when evaluating
rules in that situation, we would pass all RuleAction's but StepUp as an
allowed_action's. The rule eval logic that waterfalls through the
actions by order will then ignore actions that are not in
allowed_action's.

We could alternatively separately group "stepup rules" and then just not
evaluate those again on the second pass through KycDecisioning, but its
not clear to me yet that that is a optimal long term design and it adds
a lot of other complexities. Maybe the Rules Engine v1 work will help
inform this. For now, I think there are nice benefits to just
maintaining 1 blob of rules per obc.

## Caveats
I will I had merged this in first:
https://github.com/onefootprint/monorepo/pull/6008 😭😭😭

## Test Plan
rust integration tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->