[FP-6510] Support third-party authentication for some tenants (#6723)
## Describe your changes
Apiture wants to not have their users re-verify their phone number
inside the Footprint modal.

We accomplish this by creating an "artificial" third-party auth event,
which means that the third-party tenant has attested that the user has
logged in with then.

This skips past the identify flow and drops the user straight into the
onboarding machine. As a side effect, the user will never actually
verify their phone number, so the phone number will remain
`is_otp_verified = false`

## Caveats
There's still more to think through here on fingerprints - this will
always create global fingerprints for a user - we most likely only want
to do this if there isn't already a portable user with this fingerprint.

Though the behavior today isn't terrible - when multiple vaults match a
fingerprint, we prefer the one with a unique fingerprint over one
without.

## Test Plan
Integration test, will test locally

Here's an example of an apiture-style playbook: skip confirm, and
third-party auth event. i added a document upload in place of where we'd
do a liveness check for apiture


https://github.com/onefootprint/monorepo/assets/21066341/329767e5-1914-4af2-b822-c4d1d18138e8