Create internal API docs page (#6783)
## Describe your changes
As we start cranking out features quickly (💪) it's becoming increasingly
more necessary for backend to be able to quickly share a link to
documentation for an internal-facing API.

This uses all the amazing work we've already done to create a beautiful
internal docs page :)

When this is landed, I'll message out an announcement :)

## Caveats
On the backend, we haven't done a great job maintaining docs for these
non-public APIs. So, consumers of our new internal docs page have to
take things with a grain of salt.

Hopefully by dogfooding our own product, we'll start both making
improvements to our docs site (that customers also get!) and improving
the hygiene of our documentation on all backend APIs, not just the ones
we give to customers.

Eventually, I'd also like to put this behind an auth gate. I'd like to
eventually have docs.onefootprint.com and dashboard.onefootprint.com
share auth, but that's a more involved project. For now, we simply don't
render any links to `/internal-api-reference` - you just have to know
the path to get to our internal docs site

## Test Plan
Locally! Check out the preview link

https://docs-git-internal-api-docs.preview.onefootprint.com/internal-api-reference#get-hosted-business