Quick fix for hyphen/apostrophe name parsing from Incode (#6768)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Fix for us to handle names with hyphens better and prevent cases like:
https://onefootprint.slack.com/archives/C05U1CAD6FQ/p1699371562870999
We currently use the MRZ/barcode name from Incode if present. When we do
this, we compare it against the OCR derived name components that Incode
provides. If the MRZ/barcode name matches the OCR derived components,
then we just use the OCR components because then we can leverage
whatever name parsing logic Incode has on their end (which presumably
uses document context or something else more intelligent than what we
can do).
If a user has a hyphen in their name, it seems that the hyphen is not
present in the MRZ extracted name so for these cases, we are not
recognizing that the MRZ and OCR names match and therefore are just
parsing names out of the MRZ name (which will lack the hyphen). This PR
changes our logic so when we compare the MRZ and OCR names, we handle
properly ignore hyphens and apostrophes.


## Caveats
<!--- Are there any caveats to this changeset? -->

our name parsing logic is hardly the most robust on earth but I think we
are capturing the vast majority of cases and can continue to codify more
edge cases like this as we see them

## Test Plan
<!--- How did you test your changes? -->
unit tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->