Alpaca auto approval rules (#6784)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

Right now, for an Alpaca playbook, if address, name, or dob do not
perfectly match from the KYC vendor results, then we stepup and collect
a doc from the user. Regardless of the outcome of the doc validation, we
fail the user with review.

This PR enables to auto-approve users post step-up when possible. If the
document passes all the usual checks (document_verified,
document_selfie_matches, etc) and the name/dob/address from the document
match what was keyed in, then we pass the user and no review needed!


## Caveats
<!--- Are there any caveats to this changeset? -->

- Right now our default document rules will hard fail (without review)
in various cases (like document_not_verified, selfie_does_not_match).
But since our precedent at the moment is to always fail with review when
a doc is uploaded for Alpaca, I thought it was best to preserve this
behavior and have reviews still be raised in these 'hard fail' cases.
Unfortunately this meant adding a bit of dynamic logic to our otherwise
static rule lists.
- There's an interesting edge case where what happens when (for eg) the
KYC vendor says dob matches but then they upload a document and the dob
on the document doesn't match what was keyed in. I'm wondering if we
could still technically pass the user in this case or what Alpaca would
expect here. For now, if anything in the doc does not match, then we
will fail with review, even if KYC vendors said it matched

## Test Plan
<!--- How did you test your changes? -->
rust integration tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->