Fix open API backcompat CI (#6816)
## Describe your changes
This was horrible.... yaml is so brittle

In #6620 we added some logic to compute the merge base. Since I wrote
`run: ` instead of `run: |`, all of the commands were merged into one
line. So, the output of the command was never set. And the github
checkout action just checked out the same commit as head

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
Testing with PR on top of this that makes backcompat.

I'll probably try to add an assertion that the versions we're comparing
are different too