[FP-6610] Add all DIs to DataCollected timeline event (#6806)
## Describe your changes
Right now, if you were to edit _just_ first name, we wouldn't make a
user timeline event at all. We'd only make a timeline event if you edit
the full CDO, first name + last name.

This is a relic of when we used to disallow _any_ vaulting operation
that wasn't a full CDO. Now, when editing data via API, we don't enforce
this anymore, so we should include finer granularity here.

So, we used to serialize `attributes` as the list of full CDOs that were
updated. Now, we serialize all CDOs that had any DI updated within them.

For example, we'll say "Name" was updated if you update any piece of it