[FP-6625] Make fingerprint searching much better (#6838)
## Describe your changes
Before this PR, our fingerprint searching algorithm was pretty messy. It
generally would just split the search query into separate words, and
then compute every possible kind of fingerprint for every word in the
query, and look for vaults that matches _any_ of these queries

Flexcar had some good feedback on that old logic: if you search for
`Elliott Forde` in the dashboard, you'll get users with either name
`Elliott` or `Forde`, not just users named `Elliott Forde`.

It's pretty hard for us to search because all fields are encrypted, and
each name field is fingerprinted separately. So, this adds some pretty
complex logic to compute, given a search string, all of the permutations
of name fields that could be queried.

This also adds some logic to support AND filters when searching by
fingerprint - something we couldn't do before but that we need for
matching on id.first_name AND id.last_name simultaneously.

## Caveats
This does remove the sub-name matching functionality, but I think we
have discretion to control this. The old way of name matching was
probably objectively worse

## Test Plan
Added lots of integration tests and some unit tests on the complex name
permutation logic