add Rule PG table (#6841)
## Describe your changes

Start of BE implementation of Rules Engine v1. Mostly following the
design outlined in
https://www.notion.so/onefootprint/Rules-Engine-v1-2bfb91d5062c44d2802a18589d6e1695?pvs=4#e04c251a8a92404db8f5735c5b1f5a68

This PR adds the `Rule` table and the `create` and `update` db
operations.


## Caveats
There is an alternate design I've been contemplating here which is two
have 2 tables: `Rule` and `RuleVersion`. `RuleVersion` would basically
be what is `Rule` in this PR and then `Rule` would then represent the
"User-Facing Rule" and its `id` would be equivalent to what is now
`rule.uuid`. Haven't yet felt a super pressing need for this, but
generally feels like it might be nice. If folks have strong thoughts
there, would love to here :)


## Test Plan
<!--- How did you test your changes? -->
db tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->