Make sure that all the permission media streams are closed before camera media streams open (#6864)
## Describe your changes
We have been getting `Camera error: NotReadableError: Could not start
video source`. This error means that some other app is using the camera.
While that may very well be the case, I noticed that the "other" app
could be our app as well because the media streams we generate when we
prompt for permission aren't always getting closed before the camera
initializes its media stream. That may happen if the `getUserMedia`
promise takes more time to resolve than `navigator.permissions.query`
does. This PR makes sure that the streams are only generated when the
permission is not allowed and stop all the streams before navigating to
the camera page.


https://github.com/onefootprint/monorepo/assets/54286440/d67a564b-f9c7-4195-b10d-27b6f5cfcaaf


<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->