RuleExpression serialization (#6877)
## Describe your changes
v1 implementation of a serialization approach for `rule_expression`.
Based on latest designs, our v1 ship is going to be that each Rule can
have multiple conditions, these conditions are all AND'd together, and a
particular condition is just a boolean expression indicating the
presence or absence of a particular risk signal.

Having both `op` and `value` is not super duper necessary but:
- One of our contending designs has the user select either a `==` or
`!=` operator and `true` or `false` value.
- Other types of operators/operands such as numbers (ie experian or
incode scores) would require both an `op` and ` value` and so its
probably nice to just have this format be homogenous
- If we continue with the design where we show `not ssn_does_not_match`
instead of `ssn_does_not_match != true` in the FE, then its trivial for
the FE to consume this serialization and make that simplification.

Later we might add new operators/operands such as risk scores but
there's not firm plan or need to support that quite yet. V2 here might
also just be to go the Stripe approach and ship a SQL-like syntax and
parser and enable arbitrary boolean expressions.

## Caveats
<!--- Are there any caveats to this changeset? -->
this is a breaking/non backwards compat change so we might generate some
errors from integration tests but nbd. i'll probably just nuke all the
existing RuleInstance rows from prod/dev PG

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->