Log information on expired sessions (#6902)
## Describe your changes
- Remove a really old trigger that's existed since day 1 that used to
delete old, expired sessions. This allows us to still see information on
tokens that have expired. I don't imagine DB bloat will be a problem for
us in the near future
- Remove also some really old logic that checks for session expiry in
the db model `Session`'s `get` util function. Instead, all callers of
this function can get a session whether or not it is expired, and
display a different error message based on whether the session doesn't
exits OR is expired
- Do the expiry check _after_ logging information about the auth
session. So even if you try to use an expired auth token,the trace will
have information on what you were trying to auth as

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->