Better mock ff client (#6908)
## Describe your changes
Right now its a bit annoying that anytime we add a new flag, we need to
add mocks for this new flag in all our various test locations. It'd be
nice to just have one helper func/abstraction that gives a mocked
FFClient that by default returns the default of any value of any flag,
and then allows specific tests to mock actual values as needed. But
Mockall works in a FIFO way so its a little more annoying to do this.

This PR adds a new `MockFFClient` which will do this for us. This allows
us to remove a lot of unnecessary feature flag mocking and means we can
add new flags without changing any test code (unless we specifically
wanna test what happens when that flag is on). The Mockall interfaces
are a bit clunky/verbose so we could also use this to expose cleaner
APIs and make tests even more concise


## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->
rust tests

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->