Run Rules Engine in sandbox (#6952)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
Right now, the Rules Engine only runs in livemode because we haven't in
general been running decisioning logic in sandbox (because we just use
the fixture decision directly). This PR makes it so we still execute our
rules even in sandbox. For sandbox, the final decision will still be the
fixture decision (even if this is inconsistent with the rules output).
Still more things to figure out design/UX wise of how folks can use
rules in sandbox, but for now its probably better to at least folks
configure and edit rules in sandbox and view the results even if they
aren't directly driving the final outcome in sandbox (for now).

A few reasons to add this:
- The `test_get_rule_set_result` test passes locally but fails in prod.
This is because in prod, the integration test tenants are FF'd into
IsDemoTenant, so we implicitly fixturize these cases and effectively
treat them as sandbox:
https://github.com/onefootprint/monorepo/blob/e280778b7cee8912b47c97789eecc71f7d1cd752/backend/components/api/core/src/decision/utils.rs#L64.
Hence, at the moment in prod the Rules Engine is not actually evaluating
for the integration test tenants
- Pytests in general are not well setup for livemode testing. The SMS
OTP verify flow is pretty flaky and we don't have a great story yet
around if/what real vendor calls we are comfortable doing from
integration tests (maybe we need to check is integration/demo tenant in
places like here too:
https://github.com/onefootprint/monorepo/blob/b3453ea14a5c2db46ed11e9e456f131b839512cb/backend/components/api/core/src/decision/vendor/make_request.rs#L238)
- This will make it easier for folks to implement the FE for Rules
because they can just go ham on our APIs in sandbox



## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->
rust/py tests
<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->