Clean up error handling in document process API (#7000)
## Describe your changes
https://onefootprint.slack.com/archives/C03DQ288QQM/p1700612232316959

In this case, the client hit POST /process, timed out, and it ended up
going through and deactivating the image because it was blurry. The
client then retried POST /process because of the timeout, and there was
no active image so this assertion failed.

This fn is called to assemble all the data we pass into the incode
machine. This assertion was basically enforcing that we don't call the
incode machine with no images - it's probably a relic of old idology
logic - the incode machine will handle no image just fine

So, just removing this assertion - the incode machine will halt and
return that the current side to collect is the same side that we're
currently on. The frontend code [properly shows an
error](https://github.com/onefootprint/monorepo/blob/master/frontend/packages/idv-elements/src/plugins/id-doc/pages/processing/processing.tsx#L56-L64)
and
[retries](https://github.com/onefootprint/monorepo/blob/master/frontend/packages/idv-elements/src/plugins/id-doc/utils/state-machine/machine.ts#L233-L237)
in this case

## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->