Show decrypted first name and last name initial using transforms (#6916)
## Describe your changes
<!--- An in-depth description of the changes made by this PR. Technical
details and possible side effects. -->
### Context: [On
Slack](https://onefootprint.slack.com/archives/C0448D68Q02/p1699651858963649)

We are showing the first name decrypted and the last name initial by
default. The last name itself would still be decrypted. This PR takes
care of that. Showing only the initial of the last name is tricky. We
took an approach of using `transforms`. BE already had a notion of
`transforms` for some time now. According to our BE wizard Elliott -
'these “transforms” are a concept we already have on the backend, mostly
for the vault proxy and occasionally for some savvy users making
decryptions via API'. Learn more about `transforms` are used in BE
[here](https://docs.onefootprint.com/internal-api-reference#post-hosted-user-vault-decrypt).

We are also receiving a new field `data` in our `Entity` thanks to
Elliott. The next step would be to get rid of `attributes`,
`decryptableAttributes`, and `decryptedAttributes` from our `Entity`
type. However, it will require a massive sweep across our dashboard
codebase. I also think the `EntityVault` should include `transforms`
which is the initial approach I took in doing this task in a [different
PR](https://github.com/onefootprint/monorepo/pull/6880). However, that
also requires a lot of changes which would be appropriate to do in a
different PR when we update the `Entity` type. This PR involves minimal
changes to get the feature out the door.



https://github.com/onefootprint/monorepo/assets/54286440/fda80730-2496-425e-aca4-7b272d5275fb



## Caveats
<!--- Are there any caveats to this changeset? -->

## Test Plan
<!--- How did you test your changes? -->

<!---
### Frontend/Design Checklist
- Can we add any unit tests? What about E2E tests?
- Do we need to notify any customers?
- Do we need to update any developer docs?
- If doing SDK changes
    - Are they backwards compatible?
    - Which NPM packages do we need to release, and when?
- Are there any performance implications? What does the lighthouse
output show?

### Screenshots / videos (attachment)
- [ ] IDV
  - [ ] Mobile & Desktop
  - [ ] Bifrost & Hosted & Handoff
  - [ ] Integration with SDKs + appearance customizations
- [ ] Dashboard
- [ ] Form
  - [ ] Different variants/fields
 -->